Add optional debug logging and refactors
Refactors extracted from incoming implementation of Tests resource
There are some functional changes, such as optional debug logging

Issue: KNTK-245
