+++ examples/data-sources/kentik-synthetics_agent/provider.tf
@@ -7,6 +7,4 @@ terraform {
   }
 }
 
-provider "kentik-synthetics" {
-  # email, token and api_url are read from KTAPI_AUTH_EMAIL, KTAPI_AUTH_TOKEN, KTAPI_URL env variables
-}
+provider "kentik-synthetics" {}
