+++ synthetics/data_source_tests.go
@@ -36,9 +36,6 @@ func dataSourceTests() *schema.Resource {
 }
 
 func dataSourceTestsRead(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
-	// use UNIX time as ID to force list update every time Terraform asks for the list
-	d.SetId(strconv.FormatInt(time.Now().Unix(), 10))
-
 	resp, httpResp, err := m.(*kentikapi.Client).SyntheticsAdminServiceApi.TestsList(ctx).Execute()
 	if err != nil {
 		return detailedDiagError("failed to read tests", err, httpResp)
@@ -58,6 +55,8 @@ func dataSourceTestsRead(ctx context.Context, d *schema.ResourceData, m interfac
 		}
 	}
 
+	// use UNIX time as ID to force list update every time Terraform asks for the list
+	d.SetId(strconv.FormatInt(time.Now().Unix(), 10))
 	return nil
 }
