+++ synthetics/format.go
@@ -1,0 +1,12 @@
+package synthetics
+
+import (
+	"time"
+)
+
+func formatTime(t *time.Time) string {
+	if t == nil {
+		return ""
+	}
+	return t.Format(time.RFC3339Nano)
+}
